Add support for fixed dropdown
We can now have a menu on the top videos toolbar that flips over the
edge of the parent container with defined height. On the other side,
this position is only defined once so if a window is resized, the
fixed-positioned dropdown will stay in place. We have a dropdown
backdrop so it's not the end of the world, but it's a compromise.